requirements.txt
blinker==1.4
click==8.0.0
colorama==0.4.4
Flask==2.0.0
Flask-JWT-Extended==4.2.1
Flask-Mail==0.9.1
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
psycopg2==2.8.6
PyJWT==2.1.0
python-dotenv==0.17.1
Werkzeug==2.0.0
